pyproject.toml
[project]
name = "enginaator-2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.136.0",
    "psycopg[binary]>=3.3.3",
    "openai-whisper>=1.1.10",
    "uvicorn[standard]>=0.44.0",
    "python-multipart>=0.0.26",
    "python-dotenv>=1.2.2",
    "pydantic>=2.13.1",
    "openai>=2.32.0",
    "websockets>=16.0",
]
